requirements.txt
# Core dependencies
streamlit>=1.29.0
pandas>=2.1.0
openpyxl>=3.1.0
requests>=2.31.0
python-dotenv>=1.0.0

# Security
cryptography>=42.0.0

# Data validation
pydantic>=2.5.0

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Code quality
black>=24.0.0
flake8>=7.0.0
mypy>=1.8.0

# Logging
colorlog>=6.8.0